requirements.txt
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.16.1
channels==0.17.3
constantly==15.1.0
daphne==0.15.0
Django==1.10.3
factory-boy==2.7.0
fake-factory==0.7.2
incremental==16.10.1
msgpack-python==0.4.8
psycopg2==2.6.2
python-dateutil==2.6.0
redis==2.10.5
requests==2.12.1
Shapely==1.5.17
six==1.10.0
Twisted==16.5.0
txaio==2.5.2
zope.interface==4.3.2
